# libexpr: add more precise position information to evaluation error messages