--- a/src/libexpr/eval.cc
+++ b/src/libexpr/eval.cc
@@ -1853,21 +1853,34 @@ void ExprOpUpdate::eval(EvalState & state, Env & env, Value & v)
 
 void ExprOpConcatLists::eval(EvalState & state, Env & env, Value & v)
 {
-    Value v1; e1->eval(state, env, v1);
-    Value v2; e2->eval(state, env, v2);
-    Value * lists[2] = { &v1, &v2 };
-    state.concatLists(v, 2, lists, pos, "while evaluating one of the elements to concatenate");
+    Value v1, v2;
+    state.evalList(env, e1, v1, pos, "in the left operand of the concat (++) operator");
+    state.evalList(env, e2, v2, pos, "in the right operand of the concat (++) operator");
+
+    state.nrListConcats++;
+
+    auto l1 = v1.listSize();
+    auto l2 = v2.listSize();
+
+    if (l1 == 0) { v = v2; return; }
+    if (l2 == 0) { v = v1; return; }
+
+    state.mkList(v, l1 + l2);
+
+    auto out = v.listElems();
+    memcpy(out, v1.listElems(), l1 * sizeof(Value *));
+    memcpy(out + l1, v2.listElems(), l2 * sizeof(Value *));
 }
 
 
-void EvalState::concatLists(Value & v, size_t nrLists, Value * * lists, const PosIdx pos, std::string_view errorCtx)
+void EvalState::concatLists(Value & v, size_t nrLists, Value * * lists, const PosIdx pos)
 {
     nrListConcats++;
 
     Value * nonEmpty = 0;
     size_t len = 0;
     for (size_t n = 0; n < nrLists; ++n) {
-        forceList(*lists[n], pos, errorCtx);
+        forceList(*lists[n], pos, "while evaluating the element at index %d of the list to concatenate", n);
         auto l = lists[n]->listSize();
         len += l;
         if (l) nonEmpty = lists[n];
@@ -2062,6 +2075,7 @@ std::optional<std::string> EvalState::tryAttrsToString(const PosIdx pos, Value &
 
     return {};
 }
+
 
 BackedStringView EvalState::coerceToString(
     const PosIdx pos,