# Motivation
<!-- Briefly explain what the change is about and why it is desirable. -->

The current evaluation error trace messages are somewhat generic. When evaluating elements of a list or attribute set, we pass a static error context indicating one of the elements is being evaluated. For example, in the `prim_catAttrs` https://github.com/NixOS/nix/blob/27f82ef4a8fc589738ef06040dad5c0e214f97aa/src/libexpr/primops.cc#L2619-L2620 and in `EvalState::concatLists` https://github.com/NixOS/nix/blob/27f82ef4a8fc589738ef06040dad5c0e214f97aa/src/libexpr/eval.cc#L1894-L1910

Since we know the position of the value we are trying to evaluate from the caller side, we can add this information to the error context to give more helpful traces. A naive approach to add this information would be to construct a formatted hint from the caller side and passing it to the `foceValue` functions. However, this means the formatted hint is built on each call even if no error occurs which has a negative performance impact. This PR attempts to workaround this problem by altering the signature of the `forceValue` functions and forwarding the format arguments to their respective callee so that the formatted hint is only constructed when an evaluation error occurs.

# Context
<!-- Provide context. Reference open issues if available. -->

Initiated from the discussion at https://github.com/NixOS/nix/pull/8398#discussion_r1206513778

Benchmarks:
I performed the benchmarks using hyperfine with the current master branch being compared to this PR.

before
```
Benchmark 1: nix eval --extra-experimental-features nix-command -f ../nixpkgs/pkgs/development/haskell-modules/hackage-packages.nix
  Time (mean ± σ):     402.8 ms ±   2.5 ms    [User: 366.8 ms, System: 36.1 ms]
  Range (min … max):   399.6 ms … 407.6 ms    20 runs

Benchmark 2: nix search --extra-experimental-features "nix-command flakes" --no-eval-cache --offline ../nixpkgs hello
  Time (mean ± σ):     17.247 s ±  0.183 s    [User: 15.192 s, System: 2.128 s]
  Range (min … max):   16.958 s … 17.538 s    20 runs
```

after
```
Benchmark 1: nix eval --extra-experimental-features nix-command -f ../nixpkgs/pkgs/development/haskell-modules/hackage-packages.nix
  Time (mean ± σ):     401.2 ms ±   3.2 ms    [User: 363.0 ms, System: 38.2 ms]
  Range (min … max):   397.2 ms … 409.5 ms    20 runs

Benchmark 2: nix search --extra-experimental-features "nix-command flakes" --no-eval-cache --offline ../nixpkgs hello
  Time (mean ± σ):     17.322 s ±  0.374 s    [User: 15.235 s, System: 2.164 s]
  Range (min … max):   16.728 s … 18.134 s    20 runs
```

<!-- Non-trivial change: Briefly outline the implementation strategy. -->

<!-- Invasive change: Discuss alternative designs or approaches you considered. -->

<!-- Large change: Provide instructions to reviewers how to read the diff. -->

# Checklist for maintainers

<!-- Contributors: please leave this as is -->

Maintainers: tick if completed or explain if not relevant

 - [x] agreed on idea
 - [x] agreed on implementation strategy
 - [x] tests, as appropriate
   - functional tests - `tests/**.sh`
   - unit tests - `src/*/tests`
   - integration tests - `tests/nixos/*`
 - [ ] documentation in the manual
 - [ ] documentation in the internal API docs
 - [x] code and comments are self-explanatory
 - [x] commit message explains why the change was made
 - [ ] new feature or incompatible change: updated release notes

# Priorities

Add :+1: to [pull requests you find important](https://github.com/NixOS/nix/pulls?q=is%3Aopen+sort%3Areactions-%2B1-desc).