@polykernel 2023-06-13
@roberth I am still contemplating whether the approach will work universally but I would like to hear your thoughts. I also wanted to ask whether API breakage for libexpr is acceptable as I had some things I wanted to implement that would change how some functions are called.