@layus 2023-06-14
It makes sense to me, and the requirement to not slow execution by assembling error contexts on the fly has been addressed. So as long as everyone is okay with the code changes, I see no problem merging this.

Now, contexts were introduced with two main goals / advantages:
1. Forcing users to provide some context when they call a function likely to fail. (casting a value to a given type, mostly)
2. Avoid the bloat of having to wrap every call in a try-catch block.
3. Avoid allocating a fmt() object on the stack, at least in "hot" functions.

So, instead of making context _that_ more complex, with varargs and templates, would you consider wrapping the call site with a try-catch block that adds the error context "manually" ?