@Ericson2314 2023-06-14
@layus Are you proposing sort of the opposite approach, where we do more `try..catch` in the caller instead of less? Sounds good to me.