@layus 2023-06-14
@Ericson2314 Yes, that is my proposition. It depends a lot on the number of future occurrences of that pattern. I have counted 8 in this PR.

Now that I have had a chance to look at it properly, I find it quite elegant, given the context. It's a proper generalization of my static contexts. It gets a chance to inline, and then benefits from the zero-overhead of throw/catch.
If properly inlined, you can even pass computed values to it, like the type of a value.

Not sure if we can bring this pattern much further, but it seems to do the job as-is, well done !

I would love to get a perf review from someone with a deeper understanding of inlining, memory usage and code-size impact.