@polykernel 2023-06-15
> So, instead of making context that more complex, with varargs and templates, would you consider wrapping the call site with a try-catch block that adds the error context "manually" ?

I did initially considered following this approach but I thought adding try-catch blocks around each call site would introduce too much clutter. I think there is also the lifetime issue since objects inside the try block get destroyed at the end of the block so one needs to nest try-catch to evaluate multiple values with different context or resort to using pointers with sequential try-catch blocks.