@layus 2023-06-20
@pennae Do ypu mind running this PR in your perf benchmark ? Alternatively, do you mind sharing the benchmark and the way you run it ? I would really like to knwo the impact of this PR on perfs.

As a side note, we should really embed some benchmarks in CI builds, and report as a graph. over time.
I know sqlite has used ~qemu~ [cachegrind](https://www.sqlite.org/cpu.html) cycles count as a reliable benchmark metric independent of the CI machine and load.