@layus 2023-06-20
@pennae Thanks for your detailed explanation. Given your benchmark, I would say the OP gives pretty good confidence that this does not decrease performance, so it's all green for me.

@Ericson2314 Not that I have any commit right though :-).
I guess the only issue concerns adding "debug" parameters to core functions, but I started it :-D.